src/test/javascript/spec/app/entities/seatalloted/seatalloted.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { SeatallotedService } from 'app/entities/seatalloted/seatalloted.service';
import { ISeatalloted, Seatalloted } from 'app/shared/model/seatalloted.model';

describe('Service Tests', () => {
    describe('Seatalloted Service', () => {
        let injector: TestBed;
        let service: SeatallotedService;
        let httpMock: HttpTestingController;
        let elemDefault: ISeatalloted;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(SeatallotedService);
            httpMock = injector.get(HttpTestingController);

            elemDefault = new Seatalloted('ID', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign({}, elemDefault);
                service
                    .find('123')
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a Seatalloted', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 'ID'
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .create(new Seatalloted(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a Seatalloted', async () => {
                const returnedFromService = Object.assign(
                    {
                        seatNumber: 'BBBBBB',
                        location: 'BBBBBB',
                        floor: 'BBBBBB',
                        vacancy: 'BBBBBB',
                        requestForApproval: 'BBBBBB',
                        pendingForApproval: 'BBBBBB'
                    },
                    elemDefault
                );

                const expected = Object.assign({}, returnedFromService);
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of Seatalloted', async () => {
                const returnedFromService = Object.assign(
                    {
                        seatNumber: 'BBBBBB',
                        location: 'BBBBBB',
                        floor: 'BBBBBB',
                        vacancy: 'BBBBBB',
                        requestForApproval: 'BBBBBB',
                        pendingForApproval: 'BBBBBB'
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a Seatalloted', async () => {
                const rxPromise = service.delete('123').subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
